Aproving methods tests
