+++ method.rb
@@ -110,81 +110,3 @@ module Enumerable
     output_array
   end
 end
-
-test_array = (3...7).to_a
-
-# -----Testing my_each method-----
-
-test_array.my_each do |item|
-  puts "my_each #{item}"
-end
-
-# -----Testing my_each_with_index method-----
-
-test_array.my_each_with_index do |item, indx|
-  item << indx % 3
-  puts "my_each_with_index #{item}"
-end
-
-# -----Testing my_select method-----
-
-x = test_array.my_select do |indx|
-  (test_array[indx] % 2).zero?
-end
-puts "my_select result is : #{x}"
-
-# -----Testing my_all? method-----
-
-x = test_array.my_all? do |indx|
-  test_array[indx] > 20
-end
-puts "my_all result is : #{x}"
-
-# -----Testing my_any? method-----
-
-x = test_array.my_any? do |indx|
-  test_array[indx] >= 6
-end
-
-puts "my_any? result is : #{x}"
-
-# -----Testing my_none? method-----
-
-x = test_array.my_none? do |indx|
-  test_array[indx] > 20
-end
-puts "my_none? result is : #{x}"
-
-# -----Testing my_count method-----
-
-x = test_array.my_count { |item| item < 2 }
-
-puts "my_count result is : #{x}"
-
-# -----Testing my_inject method-----
-
-x = test_array.my_inject do |m, n|
-  (m * n)
-end
-
-puts "my_inject result is : #{x}"
-
-# -----Testing my_inject method with multiply_els method---
-
-def multiply_els(array)
-  array.my_inject do |m, n|
-    (m * n)
-  end
-end
-x = multiply_els(test_array)
-puts "Testing my_inject method with multiply_els method. The result is : #{x}"
-
-# -----Testing my_map method-----
-x = test_array.my_map do |item|
-  (item * 10)
-end
-puts "my_map with block result is : #{x}"
-
-test_proc = proc { |item| item * 20 }
-x = test_array.my_map(&test_proc)
-puts "my_map with proc result is : #{x}"
